splitted the Converge and Storage classes into seperated classes
