+++ src/Haipa.VmConfig.Primitives/VirtualMachineNetworkAdapterInfo.cs
