+++ src/Haipa.VmConfig.Primitives/VirtualMachineNetworkInfo.cs
