+++ src/Haipa.VmManagement/Haipa.VmManagement.csproj
@@ -9,6 +9,7 @@
     <PackageReference Include="Haipa.CloudInit.ConfigDrive" Version="0.5.0" />
     <PackageReference Include="Haipa.CloudInit.ConfigDrive.NoCloud" Version="0.5.0" />
     <PackageReference Include="Haipa.CloudInit.ConfigDrive.WindowsImaging" Version="0.5.0" />
+    <PackageReference Include="IPNetwork2" Version="2.1.2" />
     <PackageReference Include="LanguageExt.Core" Version="3.4.15" />
     <PackageReference Include="Microsoft.Windows.Compatibility" Version="3.1.1" />
     <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
