+++ src/Haipa.VmManagement/Storage/VMDVdStorageSettings.cs
@@ -1,0 +1,7 @@
+﻿namespace Haipa.VmManagement.Storage
+{
+    public class VMDvDStorageSettings : VMDriveStorageSettings
+    {
+
+    }
+}
